MGS-7964 gpu: Integrate Mali DDK r53p0
- Added ioctl to query tiler heap size and peak size.

- Move CS_USER IO allocation from CS binding to CSG creation.

- Add reserved field to QUEUE_GROUP_CREATE ioctl for future use.
- Previous version retained as KBASE_IOCTL_CS_QUEUE_GROUP_CREATE_1_35
  for backwards compatibility.

+++ drivers/gpu/arm/midgard/csf/mali_kbase_csf_fw_io.h
@@ -320,7 +320,7 @@ void kbase_csf_fw_io_group_write64(struct kbase_csf_fw_io *fw_io, u32 group_id, 
  *
  * Return: Value of the word read from a CSG's input page.
  */
-u32 kbase_csf_fw_io_group_input64_read(struct kbase_csf_fw_io *fw_io, u32 group_id, u32 offset);
+u64 kbase_csf_fw_io_group_input64_read(struct kbase_csf_fw_io *fw_io, u32 group_id, u32 offset);
 
 /**
  * kbase_csf_fw_io_group_write64_mask() - Write part of a double-word in a CSG's input page.
